flask-api/requirements.txt
alembic==1.5.2
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
linecache2==1.0.0
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
traceback2==1.4.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
